+++ src/Part.java
@@ -5,8 +5,8 @@ import java.awt.Color;
 
 public class Part {
 	GameControler controler;
-    double x, y, w, h;
-    double xv, yv;
+    float x, y, w, h;
+    float xv, yv;
     Color color;
     Random rand;
     boolean alive;
@@ -20,18 +20,18 @@ public class Part {
 
 	public void paint(Graphics g) {}
     public void update() {}
-    public double getX() { return x; }
-    public void setX(double x) { this.x = x; }
-    public double getY() { return y; }
-    public void setY(double x) { this.y = x; }
-    public double getWidth() { return w; }
-    public void setWidth(double x) { this.w = x; }
-    public double getHeight() { return h; }
-    public void setHeight(double x) { this.h = x; }
-    public double getXV() { return xv; }
-    public void setXV(double x) { xv = x; }
-    public double getYV() { return yv; }
-    public void setYV(double y) { yv = y; }
+    public float getX() { return x; }
+    public void setX(float x) { this.x = x; }
+    public float getY() { return y; }
+    public void setY(float x) { this.y = x; }
+    public float getWidth() { return w; }
+    public void setWidth(float x) { this.w = x; }
+    public float getHeight() { return h; }
+    public void setHeight(float x) { this.h = x; }
+    public float getXV() { return xv; }
+    public void setXV(float x) { xv = x; }
+    public float getYV() { return yv; }
+    public void setYV(float y) { yv = y; }
     public boolean isAlive() { return alive; }
     public boolean collides(Player p) { return false; };
     public boolean collides(Bullet b) { return false; };
