+++ src/Player.java
@@ -9,7 +9,7 @@ public class Player extends Part {
     private int SHOOTING_DELAY = 5;
 
     Bullet nextBullet;
-    double angle;
+    float angle;
     int hit;
     boolean showHit;
 
@@ -101,12 +101,12 @@ public class Player extends Part {
    
     public void mouseMoved(MouseEvent e) {
         if (hit > 0) return;
-        double dy = y - e.getY();
-        double dx = x - e.getX();
+        float dy = y - e.getY();
+        float dx = x - e.getX();
 
         if (dx <= 0) return;
 
         if (dy == 0) angle = 0;
-        else angle = Math.atan(dy / dx);
+        else angle = (float) Math.atan(dy / dx);
     }
 }
